+++ ValheimCommunityPatch/ValheimCommunityPatch.cs
@@ -18,7 +18,7 @@ namespace ValheimCommunityPatch
     {
         public const string PluginGUID = "MidnightsFX.ValheimCommunityPatch";
         public const string PluginName = "ValheimCommunityPatch";
-        public const string PluginVersion = "0.26.0";
+        public const string PluginVersion = "0.26.1";
 
         internal static ManualLogSource Log;